[CWS] shift a u64 when building and testing event masks
The auid, basename and in_upper_layer approvers select an event with
`event_mask & (1 << (event_type - 1))`, and add_event_to_mask sets a bit
in a discarder mask the same way. Both operate on a u64 while shifting
an int, so only the low 32 bits are reachable.

Every event type that reaches these paths is at most EVENT_CHDIR, which
is 15, so they are correct today. They are one renumbering or one new
file event away from silently dropping events, and the mask is already
large enough to hold the 69 kernel event types. add_event_to_mask is the
sharper case: the mask it writes is read back by mask_has_event, which
already shifts a u64, so the pair disagrees above bit 31.

Shift a u64 in all four, matching mask_has_event and flag_approver.

+++ pkg/security/ebpf/c/include/helpers/approvers.h
@@ -256,7 +256,7 @@ static enum SYSCALL_STATE __attribute__((always_inline)) approve_by_auid(struct 
     u32 auid = pid_entry->credentials.auid;
 
     struct event_mask_filter_t *mask_filter = bpf_map_lookup_elem(&auid_approvers, &auid);
-    if (mask_filter && mask_filter->event_mask & (1 << (event_type - 1))) {
+    if (mask_filter && mask_filter->event_mask & ((u64)1 << (event_type - 1))) {
         monitor_event_approved(syscall->type, AUID_APPROVER_TYPE);
         return ACCEPTED;
     }
@@ -285,7 +285,7 @@ static enum SYSCALL_STATE __attribute__((always_inline)) flag_approver (struct u
 
 static int __attribute__((always_inline)) is_basename_in_map(struct basename_t *basename, u64 event_type) {
     struct event_mask_filter_t *filter = bpf_map_lookup_elem(&basename_approvers, basename);
-    return filter && filter->event_mask & (1 << (event_type - 1));
+    return filter && filter->event_mask & ((u64)1 << (event_type - 1));
 }
 
 static enum SYSCALL_STATE __attribute__((always_inline)) approve_by_basename(struct dentry *dentry, u64 event_type) {
@@ -327,7 +327,7 @@ static enum SYSCALL_STATE __attribute__((always_inline)) approve_by_basename(str
 static enum SYSCALL_STATE __attribute__((always_inline)) approve_by_in_upper_layer(u64 event_type, struct file_t *file) {
     u32 key = 0;
     struct event_mask_filter_t *filter = bpf_map_lookup_elem(&in_upper_layer_approvers, &key);
-    if (filter && filter->event_mask & (1 << (event_type - 1)) && (file->flags & UPPER_LAYER) > 0) {
+    if (filter && filter->event_mask & ((u64)1 << (event_type - 1)) && (file->flags & UPPER_LAYER) > 0) {
         monitor_event_approved(event_type, IN_UPPER_LAYER_APPROVER_TYPE);
         return APPROVED;
     }
